+++ odps-sdk-impl/odps-udf-local/src/main/java/com/aliyun/odps/udf/local/LocalExecutionContext.java
@@ -149,4 +149,9 @@ public class LocalExecutionContext extends ExecutionContext {
     return null;
   }
 
+  //TODO
+  @Override
+  public FileSystem getTempFileSystem() throws IOException {
+    return null;
+  }
 }