+++ odps-sdk/odps-sdk-udf/src/main/java/com/aliyun/odps/udf/ExecutionContext.java
@@ -274,4 +274,6 @@ public abstract class ExecutionContext {
   public abstract FileSystem getOutputVolumeFileSystem(String label) throws IOException;
 
   public abstract Properties getConfigurations();
+
+  public abstract FileSystem getTempFileSystem() throws IOException;
 }